Refactor to two components
